Rename legacy visual file paths
