+++ actions/capture/action.yml
@@ -52,8 +52,8 @@ runs:
         import path from 'node:path';
         import { pathToFileURL } from 'node:url';
 
-        const modulePath = path.resolve(process.env.ACTION_ROOT, 'lib/capture-visual-routes.mjs');
-        const configModulePath = path.resolve(process.env.ACTION_ROOT, 'lib/visual-regression-config.mjs');
+        const modulePath = path.resolve(process.env.ACTION_ROOT, 'lib/capture-routes.mjs');
+        const configModulePath = path.resolve(process.env.ACTION_ROOT, 'lib/snapdrift-config.mjs');
         const { runBaselineCapture } = await import(pathToFileURL(modulePath));
         const { splitCommaList } = await import(pathToFileURL(configModulePath));
