+++ actions/compare/action.yml
@@ -116,8 +116,8 @@ runs:
         import path from 'node:path';
         import { pathToFileURL } from 'node:url';
 
-        const modulePath = path.resolve(process.env.ACTION_ROOT, 'lib/compare-visual-results.mjs');
-        const configModulePath = path.resolve(process.env.ACTION_ROOT, 'lib/visual-regression-config.mjs');
+        const modulePath = path.resolve(process.env.ACTION_ROOT, 'lib/compare-results.mjs');
+        const configModulePath = path.resolve(process.env.ACTION_ROOT, 'lib/snapdrift-config.mjs');
         const { runDriftCheckCli } = await import(pathToFileURL(modulePath));
         const { splitCommaList } = await import(pathToFileURL(configModulePath));
