+++ lib/capture-routes.mjs
@@ -14,7 +14,7 @@ import {
   SNAPDRIFT_VIEWPORT_PRESETS,
   SNAPDRIFT_NAVIGATION_TIMEOUT_MS,
   SNAPDRIFT_SETTLE_DELAY_MS
-} from './visual-regression-config.mjs';
+} from './snapdrift-config.mjs';
 
 /** @typedef {import('../types/visual-diff-types').VisualBaselineResults} BaselineResults */
 /** @typedef {import('../types/visual-diff-types').VisualBaselineRouteResult} BaselineRouteResult */
