+++ lib/compare-results.mjs
@@ -11,7 +11,7 @@ import {
   resolveFromWorkingDirectory,
   selectConfiguredRoutes,
   splitCommaList
-} from './visual-regression-config.mjs';
+} from './snapdrift-config.mjs';
 
 const { PNG } = pngjs;
