+++ lib/drift-summary.mjs
@@ -3,7 +3,7 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
 
-import { splitCommaList } from './visual-regression-config.mjs';
+import { splitCommaList } from './snapdrift-config.mjs';
 
 const DEFAULT_OUT_DIR = path.join('qa-artifacts', 'snapdrift', 'drift', 'current');
