+++ lib/stage-artifacts.mjs
